- Fixed a render bug in edit mode. - Collision detection is currently deactivated for most objects (for testing purposes, going to rewrite   most of it) - Some method prototypes were changed to remove some obsolete stuff. - Fixed the problem that units (mainly the stank) could get locked up in walls of the garage. - Some minor other cleanups and fixes. - Ignore all those vssver.scc files, they are from my local VSS database and contain some bookkeeping info   specifically for my machine.
git-svn-id: http://192.168.1.80/svn/golgotha/worksrc@115 dfb5a597-25cf-0310-bb4d-ec683fa0f420

+++ editor__solvemap_astar2.cpp
@@ -370,7 +370,7 @@ void Path::GenerateSucc(struct NODE *BestNode, int x, int y, int dx, int dy)
         //Successor->h = MAX(deltaX, deltaY);//maximum distance extimate
 		//Successor->h=sqrt(deltaX*deltaX+deltaY*deltaY);//euclidean (would require floats)
 		//Successor->h=(deltaX==deltaY)?sqrt2:1;//faster euclidean(dito)
-		Successor->h=deltaX+deltaY;//Manhattan distance extimate
+		Successor->h=(float)deltaX+deltaY;//Manhattan distance extimate
 
         Successor->f = g + Successor->h;
         Successor->x = x;
